# optimize(Packages): use mmap to read large files