Using less syscalls and processing data in memory makes getNumStringsImpl ~400% faster.

Some benchmark highlights:

On my Ubuntu 22.04 laptop, with 1408 dpkg packages.

```text
Run on (8 X 2400.01 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.12, 0.06, 0.01
---------------------------------------------------------------------
Benchmark                           Time             CPU   Iterations
---------------------------------------------------------------------
BM_getNumStringsImpl           256691 ns       256694 ns         2517
BM_getNumStringsLegacyImpl    1080674 ns      1080676 ns          636
```