@CarterLi 2023-12-07
Can't we use FF_AUTO_CLOSE_FD?