@CarterLi 2023-12-07
Also, all these early returns should fall back to the old function