@apocelipes 2023-12-07
> Also, all these early returns should fall back to the old function

I don't think fdopen can work if fstat already failed.

mmap is a bit complicated, and there are a few errors where it can fall back. This kind of situation is quite rare. The only thing that is more likely to happen is mmap failure due to no enough memory. Is it really worth falling back here?