@CarterLi 2023-12-07
I am actually not familiar with mmap.

Did you test ffReadFileBuffer? How does ffReadFileBuffer compare to mmap?