@CarterLi 2023-12-07
Also, you should use benchmark fastfetch itself, not pull things out. I suggest hyperfine. If the benchmark result differs not much, I prefer use the simpler implementation.