@apocelipes 2023-12-07
> I am actually not familiar with mmap.
> 
> Did you test ffReadFileBuffer? How does ffReadFileBuffer compare to mmap?

BM_getNumStringsImpl                    1400988 ns      1400979 ns          476
BM_getNumStringsLegacyImpl              6468152 ns      6468036 ns          108
BM_getNumStringsWithffReadFileBuffer    1839101 ns      1839052 ns          376

Seems like 30% slower than mmap.

the code:

```c
static uint32_t useFFReadBuffer(const char* filename, const char* needle)
{
    FF_STRBUF_AUTO_DESTROY content = ffStrbufCreate();
    if (!ffReadFileBuffer(filename, &content))
        return 0;
    uint32_t count = 0;
    size_t needleLength = strlen(needle);
    char *iter = content.chars;
    while ((iter = strstr(iter, needle)) != NULL)
    {
	++count;
        iter += needleLength;
    }
    return count;
}
```