@CarterLi 2023-12-07
What about ReadFileBuffer?