@apocelipes 2023-12-07
> What about ReadFileBuffer?

single module: 
![image](https://github.com/fastfetch-cli/fastfetch/assets/21255940/1ee34db0-8df4-4581-a2e7-fea53bbe3bfe)

![image](https://github.com/fastfetch-cli/fastfetch/assets/21255940/913f27f6-2420-4b47-8b82-9859c6cda2b7)

entire command:

![image](https://github.com/fastfetch-cli/fastfetch/assets/21255940/1ea631bf-b8b3-473a-b5a6-a4708956db33)

Not as much fast as mmap, but is still faster than stdio.