@CarterLi 2023-12-07
You can use `hyperfine './ff-stdio -s package' './ff-buffer -s package' './ff-mmap -s package'`

I prefer `ff-buffer` since its implementation is the simplest.