@apocelipes 2023-12-07
> You can use `hyperfine './ff-stdio -s package' './ff-buffer -s package' './ff-mmap -s package'`
> 
> I prefer `ff-buffer` since its implementation is the simplest.

![image](https://github.com/fastfetch-cli/fastfetch/assets/21255940/979f8697-88a5-4b69-99a2-6c96a45a47ea)

Faster than mmap may be an inaccuracy.

By the way, my benchmarking environment only has 1410 packages, whether other distros could have 2000+ packages. Theoretically the more packages you have the more significant the performance difference.

If u prefer ffReadBuffer, I'll change to it later, and getNumStringsLegacyImpl can be deleted.