@CarterLi 2023-12-07
> If u prefer ffReadBuffer, I'll change to it later, and getNumStringsLegacyImpl can be deleted

Please do it.

And you may use `memmem` instead of `strstr`, since lengths of both string are known.